changed estimatedTimeEnRoute from int to long

+++ src/org/dinopolis/gpstool/gpsinput/garmin/GarminWaypoint.java
@@ -229,7 +229,7 @@ public interface GarminWaypoint
  * @throws UnsupportedOperationException
  */
 	
-	public int getEstimatedTimeEnroute()
+	public long getEstimatedTimeEnroute()
     throws UnsupportedOperationException;
 
 //----------------------------------------------------------------------
